+++ Makefile
@@ -25,6 +25,8 @@ help: ## Display this help.
 .PHONY: generate
 generate: gen-proto
 	go generate ./...
+	cp pkg/bpmn/persistence/rqlite/sql/db.go.template pkg/bpmn/persistence/rqlite/sql/db.go
+	sed -i "s/Foreign[[:space:]]\+interface{}[[:space:]]\+\`json:\"foreign\"\`//g" pkg/bpmn/persistence/rqlite/sql/models.go
 
 .PHONY: gen-proto
 gen-proto:
