+++ internal/rqlite/sql/after-generate.sh
