+++ internal/rqlite/sql/db.go.template
