+++ internal/rqlite/sql/generate.go
@@ -1,4 +1,3 @@
 package sql
 
 //go:generate sqlc generate
-//go:generate ./after-generate.sh
