+++ internal/rqlite/sql/message_subscription.sql.go
