+++ internal/rqlite/sql/sqlc.yaml
