+++ internal/rqlite/sql_source/activity_instance.sql
