+++ internal/rqlite/sql_source/message_subscription.sql
