+++ internal/rqlite/sql_source/timer.sql
