+++ pkg/bpmn/persistence_service.go
@@ -1,6 +1,6 @@
 package bpmn
 
-import rqlite "github.com/pbinitiative/zenbpm/pkg/bpmn/persistence/rqlite"
+import rqlite "github.com/pbinitiative/zenbpm/internal/rqlite"
 
 type BpmnEnginePersistenceService interface {
 	FindProcessById(processId string) *ProcessInfo
